refactor(web): split handlers into domain modules and layered services
Decompose monolithic web handlers into focused auth/forum/moderation/OAuth modules backed by service-repository layers, and load OAuth credentials from environment variables to keep secrets out of source.

Made-with: Cursor
