+++ cmd/web/middleware.go
@@ -1,88 +1,35 @@
 package main
 
 import (
-	"fmt"
 	"net/http"
-	"sync"
 	"time"
+
+	mw "github.com/aspandyar/forum/internal/transport/http/middleware"
 )
 
 var (
-	requestCount   = make(map[string]int)
-	requestCountMu sync.Mutex
-	limit          = 100
-	period         = time.Minute
+	limit  = 100
+	period = time.Minute
 )
 
+var rateLimiter = mw.RateLimit(limit, period)
+
 func rateLimitMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		clientIP := r.RemoteAddr // You may need to extract the client's IP more accurately
-
-		requestCountMu.Lock()
-		defer requestCountMu.Unlock()
-
-		// Check if the client has exceeded the rate limit
-		if requestCount[clientIP] >= limit {
-			http.Error(w, "Rate limit exceeded", http.StatusTooManyRequests)
-			return
-		}
-
-		// Increment the request count for this client
-		requestCount[clientIP]++
-
-		// Reset the request count after the specified period
-		time.AfterFunc(period, func() {
-			requestCountMu.Lock()
-			defer requestCountMu.Unlock()
-			requestCount[clientIP] = 0
-		})
-
-		next.ServeHTTP(w, r)
-	})
+	return rateLimiter(next)
 }
 
 func secureHeaders(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// w.Header().Set("Content-Security-Policy", "default-src 'self'; frame-src 'self' https://accounts.google.com;  style-src 'self' fonts.googleapis.com; font-src 'self' fonts.googleapis.com;")
-
-		w.Header().Set("Referrer-Policy", "origin-when-cross-origin")
-		w.Header().Set("X-Content-Type-Options", "nosniff")
-		w.Header().Set("X-Frame-Options", "deny")
-		w.Header().Set("X-XSS-Protection", "0")
-
-		next.ServeHTTP(w, r)
-	})
+	return mw.SecureHeaders(next)
 }
 
 func (app *application) logRequest(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		app.infoLog.Printf("%s - %s %s %s", r.RemoteAddr, r.Proto, r.Method, r.URL.RequestURI())
-
-		next.ServeHTTP(w, r)
-	})
+	return mw.LogRequest(app.infoLog)(next)
 }
 
 func (app *application) recoverPanic(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		defer func() {
-			if err := recover(); err != nil {
-				w.Header().Set("Connection", "close")
-
-				app.serverError(w, fmt.Errorf("%s", err))
-			}
-		}()
-
-		next.ServeHTTP(w, r)
-	})
+	return mw.RecoverPanic(app.serverError)(next)
 }
 
 func (app *application) requireAuthentication(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if !app.isAuthenticated(r) {
-			http.Redirect(w, r, "/user/login", http.StatusSeeOther)
-			return
-		}
-		w.Header().Add("Cache-Control", "no-store")
-		next.ServeHTTP(w, r)
-	})
+	return mw.RequireAuthentication(app.isAuthenticated)(next)
 }
