+++ cmd/web/template.go
@@ -1,59 +1,13 @@
 package main
 
 import (
-	"path/filepath"
 	"text/template"
-	"time"
 
-	"github.com/aspandyar/forum/internal/models"
+	renderpkg "github.com/aspandyar/forum/internal/transport/http/render"
 )
 
-type templateData struct {
-	CurrentYear     int
-	Forum           *models.Forum
-	Forums          []*models.Forum
-	Form            interface{}
-	Flash           string
-	IsAuthenticated bool
-	Role            int
-}
-
-func humanDate(t time.Time) string {
-	return t.Format("02 Jan 2006 at 15:04")
-}
-
-var functions = template.FuncMap{
-	"humanDate": humanDate,
-}
+type templateData = renderpkg.TemplateData
 
 func newTemplateCache() (map[string]*template.Template, error) {
-	cache := map[string]*template.Template{}
-
-	pages, err := filepath.Glob("./ui/html/pages/*.tmpl.html")
-	if err != nil {
-		return nil, err
-	}
-
-	for _, page := range pages {
-		name := filepath.Base(page)
-
-		ts, err := template.New(name).Funcs(functions).ParseFiles("./ui/html/base.tmpl.html")
-		if err != nil {
-			return nil, err
-		}
-
-		ts, err = ts.ParseGlob("./ui/html/partials/*.tmpl.html")
-		if err != nil {
-			return nil, err
-		}
-
-		ts, err = ts.ParseFiles(page)
-		if err != nil {
-			return nil, err
-		}
-
-		cache[name] = ts
-	}
-
-	return cache, nil
+	return renderpkg.NewTemplateCache()
 }
