+++ internal/transport/http/middleware/middleware.go
@@ -1,0 +1,84 @@
+package middleware
+
+import (
+	"fmt"
+	"log"
+	"net/http"
+	"sync"
+	"time"
+)
+
+func RateLimit(limit int, period time.Duration) func(http.Handler) http.Handler {
+	requestCount := make(map[string]int)
+	var requestCountMu sync.Mutex
+
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			clientIP := r.RemoteAddr
+
+			requestCountMu.Lock()
+			defer requestCountMu.Unlock()
+
+			if requestCount[clientIP] >= limit {
+				http.Error(w, "Rate limit exceeded", http.StatusTooManyRequests)
+				return
+			}
+
+			requestCount[clientIP]++
+
+			time.AfterFunc(period, func() {
+				requestCountMu.Lock()
+				defer requestCountMu.Unlock()
+				requestCount[clientIP] = 0
+			})
+
+			next.ServeHTTP(w, r)
+		})
+	}
+}
+
+func SecureHeaders(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Referrer-Policy", "origin-when-cross-origin")
+		w.Header().Set("X-Content-Type-Options", "nosniff")
+		w.Header().Set("X-Frame-Options", "deny")
+		w.Header().Set("X-XSS-Protection", "0")
+		next.ServeHTTP(w, r)
+	})
+}
+
+func LogRequest(logger *log.Logger) func(http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			logger.Printf("%s - %s %s %s", r.RemoteAddr, r.Proto, r.Method, r.URL.RequestURI())
+			next.ServeHTTP(w, r)
+		})
+	}
+}
+
+func RecoverPanic(serverError func(http.ResponseWriter, error)) func(http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			defer func() {
+				if err := recover(); err != nil {
+					w.Header().Set("Connection", "close")
+					serverError(w, fmt.Errorf("%s", err))
+				}
+			}()
+			next.ServeHTTP(w, r)
+		})
+	}
+}
+
+func RequireAuthentication(isAuthenticated func(*http.Request) bool) func(http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			if !isAuthenticated(r) {
+				http.Redirect(w, r, "/user/login", http.StatusSeeOther)
+				return
+			}
+			w.Header().Add("Cache-Control", "no-store")
+			next.ServeHTTP(w, r)
+		})
+	}
+}
